FEAT: Novas camadas em estoque, controller, service, etc

+++ src/main/java/com/talentoemlinha/controller/FuncionarioController.java
@@ -13,10 +13,7 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
-// import com.talentoemlinha.dto.PontoDto;
 import com.talentoemlinha.model.Funcionario;
-// import com.talentoemlinha.model.Ponto;
-// import com.talentoemlinha.service.PontoService;
 import com.talentoemlinha.service.FuncionarioService;
 
 @RestController
@@ -24,7 +21,6 @@ public class FuncionarioController {
 
     @Autowired
     private FuncionarioService funcService;
-    // private PontoService pontoService;
 
     @GetMapping("/funcionario")
     public ResponseEntity<List<Funcionario>> funcionarioGet() {
@@ -32,30 +28,13 @@ public class FuncionarioController {
     }
 
     @GetMapping("/funcionario/{id}")
-    public ResponseEntity<Funcionario> funcionarioGet(@PathVariable int id) {
+    public ResponseEntity<Funcionario> funcionarioGet(@PathVariable long id) {
         Funcionario responseFunc = funcService.retornarFuncionarioPeloId(id);
         if (responseFunc != null)
             return ResponseEntity.status(HttpStatus.OK).body(responseFunc);
         return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
 
     }
-
-    // @PostMapping("/funcionario/{id}/bonificar")
-    // public ResponseEntity<Ponto> funcionarioPost(@PathVariable int id,
-    // @RequestBody PontoDto pontos) {
-    // for (Funcionario funcionario : funcionarios) {
-    // if (funcionario.getNp() == id){
-    // int indicePonto = pontoService.retornarTodosPontos().size()+1;
-    // Ponto novoPonto = new
-    // Ponto(indicePonto,id,pontos.getQuantidade(),pontos.getMotivo());
-    // funcionario.setTotalDePontos(funcionario.getTotalDePontos() +
-    // pontos.getQuantidade());
-    // return
-    // ResponseEntity.status(HttpStatus.CREATED).body(pontoService.adicionarPonto(novoPonto));
-    // }
-    // }
-    // return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
-    // }
 
     @PostMapping("/funcionario")
     public ResponseEntity<Funcionario> funcionarioPost(@RequestBody Funcionario func) {
@@ -66,7 +45,7 @@ public class FuncionarioController {
     }
 
     @PutMapping("/funcionario/{id}")
-    public ResponseEntity<Funcionario> funcionarioPut(@PathVariable int id, @RequestBody Funcionario func) {
+    public ResponseEntity<Funcionario> funcionarioPut(@PathVariable long id, @RequestBody Funcionario func) {
         Funcionario resposeFunc = funcService.alterarFuncionario(id, func);
         if (resposeFunc == null)
             return null;
@@ -74,7 +53,7 @@ public class FuncionarioController {
     }
 
     @DeleteMapping("/funcionario/{id}")
-    public ResponseEntity<Funcionario> funcionarioDelete(@PathVariable int id) {
+    public ResponseEntity<Funcionario> funcionarioDelete(@PathVariable long id) {
         if (funcService.deletarFuncionario(id) != null)
             return ResponseEntity.status(HttpStatus.ACCEPTED).body(null);
         return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
